requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
torch==2.2.2
torchvision==0.17.2
transformers==4.40.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4